Practiced array_splice() for a while. It is a slightly confusing function it
mutates the array and returns the slice. The "offset" parameter is actually the index of the array, and the slice is from "offset" to "length" inclusive if both offset and length are positives integers. Using negative offsets or lengths is a bit more confusing and will require more practice to better comprehend